requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.3.0
Pillow==10.1.0
drf-yasg==1.21.7
django-filter==23.5
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
